Mission 2.3 : Gestion des niveaux (ajout, suprression)
